+++ datastream-server/src/main/java/com/linkedin/datastream/server/PartitionThroughputInfo.java
@@ -5,15 +5,15 @@ package com.linkedin.datastream.server;
  * A structure for partition throughput information.
  */
 public class PartitionThroughputInfo {
-  private final Long _bytesInRate;
-  private final Long _messagesInRate;
+  private final int _bytesInRate;
+  private final int _messagesInRate;
 
   /**
    * Creates an instance of {@link PartitionThroughputInfo}
    * @param bytesInRate Bytes in rate for the partition
    * @param messagesInRate Messages in rate for the partition
    */
-  public PartitionThroughputInfo(Long bytesInRate, Long messagesInRate) {
+  public PartitionThroughputInfo(int bytesInRate, int messagesInRate) {
     _bytesInRate = bytesInRate;
     _messagesInRate = messagesInRate;
   }
@@ -21,14 +21,14 @@ public class PartitionThroughputInfo {
   /**
    * Gets the bytes in rate
    */
-  public Long getBytesInRate() {
+  public int getBytesInRate() {
     return _bytesInRate;
   }
 
   /**
    * Gets the messages in rate
    */
-  public Long getMessagesInRate() {
+  public int getMessagesInRate() {
     return _messagesInRate;
   }
 }
